rag_api/requirements.txt
langchain>=0.0.350
chromadb>=0.4.22
sentence-transformers>=2.2.2
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
openai>=1.12.0
tiktoken>=0.5.2
boto3>=1.34.0
python-magic>=0.4.27
python-magic-bin>=0.4.14
unstructured>=0.10.30
pdf2image>=1.16.3
pytesseract>=0.3.10
pinecone-client>=2.2.4
weaviate-client>=3.24.1
qdrant-client>=1.7.0
pydantic-settings>=2.1.0 